+++ react/Tooltip.js
@@ -1,0 +1,3 @@
+import Tooltip from './components/Toggle/index'
+
+export default Tooltip